requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pygithub
aiofiles